+++ index.html
@@ -3,11 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script
-      defer
-      src="https://kit.fontawesome.com/b0c0f87c06.js"
-      crossorigin="anonymous"
-    ></script>
+    <link rel="icon" type="public/ico" href="favicon.ico" />
+    <link rel="stylesheet" type="text/css" href="https://necolas.github.io/normalize.css/8.0.1/normalize.css" />
+    <script defer src="https://kit.fontawesome.com/b0c0f87c06.js" crossorigin="anonymous"></script>
     <title>quiz</title>
   </head>
   <body>
