+++ src/main.css
@@ -26,7 +26,8 @@ body {
 .header {
   background-color: var(--main-dark);
   color: var(--text-light);
-  padding: 30px;
+  height: 110px;
+
   display: flex;
   gap: 30px;
   flex-direction: column;
@@ -36,6 +37,9 @@ body {
   background-color: var(--main-light);
   padding: 30px;
   min-height: 600px;
+}
+h2 {
+  color: var(--main-dark);
 }
 
 .buttons-wrapper {
@@ -137,6 +141,16 @@ h4 {
   border-right: 2px solid var(--main-dark);
   animation: none;
 }
+
+.error-page {
+  height: 80vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 30px;
+}
+
 @keyframes spinLeft {
   from {
     transform: rotate(0deg);
