--- src/pages/ErrorPage.css
@@ -1,7 +1,0 @@
-.error-page {
-  height: 80vh;
-  color: var(--text-light);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
