+++ src/pages/ErrorPage/index.tsx
@@ -1,5 +1,4 @@
 import { Link, useRouteError } from 'react-router-dom';
-import './ErrorPage.css';
 
 const ErrorPage = () => {
   const error = useRouteError();
@@ -9,9 +8,11 @@ const ErrorPage = () => {
     <div className="error-page">
       <h1>Oops!</h1>
       <p>Sorry, an unexpected error has occurred.</p>
-      <Link to="/" className={'back-btn'}>
-        Back to settings
-      </Link>
+      <p>
+        <Link to="/" className={'back-btn'}>
+          Back to settings
+        </Link>
+      </p>
     </div>
   );
 };
