+++ src/pages/ResultPage/index.css
