+++ src/pages/StatisticsPage/index.css
